Ensure sentry is treeshaken
