+++ src/utils/errors.js
@@ -1,0 +1,40 @@
+let Sentry
+
+export function configureSentry (app, { dsn, environment }) {
+  if (process.env.DEV) {
+    if (process.env.ENABLE_DEV_SENTRY !== 'true') {
+      return
+    }
+    environment = 'dev'
+  }
+
+  // Async, import so we only load sentry lib if actually using it
+  import('./sentry').then(({ init, captureMessage, captureException }) => {
+    Sentry = { captureMessage, captureException }
+    init({
+      app,
+      dsn,
+      environment,
+      release: process.env.KARROT.GIT_SHA1,
+
+      // Disable sentry session tracking
+      // See https://docs.sentry.io/platforms/javascript/configuration/releases/?original_referrer=https%3A%2F%2Fwww.startpage.com%2F#sessions
+      // and https://community.karrot.world/t/sending-too-much-data-to-sentry/1194
+      autoSessionTracking: false,
+      ignoreErrors: [
+        'ResizeObserver loop limit exceeded', // Chrome
+        'ResizeObserver loop completed with undelivered notifications', // Firefox
+      ],
+    })
+  })
+}
+
+export function captureException (...args) {
+  if (!Sentry) return
+  Sentry.captureException(...args)
+}
+
+export function captureMessage (...args) {
+  if (!Sentry) return
+  Sentry.captureMessage(...args)
+}
